Fix wrong encoding image (commit 58e5bafffe828ac1891a19f57b0467fb8a67f8f8).

+++ metapath/src/Helpers.c
@@ -430,7 +430,7 @@ void SnapToDefaultButton(HWND hwndBox) {
 //
 //  MakeBitmapButton()
 //
-void MakeBitmapButton(HWND hwnd, int nCtlId, HINSTANCE hInstance, UINT uBmpId) {
+void MakeBitmapButton(HWND hwnd, int nCtlId, HINSTANCE hInstance, WORD uBmpId) {
 	HWND hwndCtl = GetDlgItem(hwnd, nCtlId);
 	HBITMAP hBmp = LoadImage(hInstance, MAKEINTRESOURCE(uBmpId), IMAGE_BITMAP, 0, 0, LR_CREATEDIBSECTION);
 	BITMAP bmp;
